Renamve event V1 converter filename
