+++ packages/notifier-seeder/src/models/catalogEventV2ConverterToV1.ts
